+++ physics/double2.h
@@ -5,7 +5,7 @@ class double2
 {
 public:
     double2() : x_(0.0), y_(0.0) {}
-    double2(double x, double y) : x_(x), y_(y) {}
+    double2(float x, float y) : x_(x), y_(y) {}
     double2& operator=(const double2& r)
     {
         if (this != &r)
@@ -35,7 +35,7 @@ public:
     {
         return double2(x_ - r.x_, y_ - r.y_);
     }
-    double2 operator/(double d) const
+    double2 operator/(float d) const
     {
         return double2(x_ / d, y_ / d);
     }
@@ -43,22 +43,22 @@ public:
     {
         return double2(-x_, -y_);
     }
-    double2 operator*(double r) const
+    double2 operator*(float r) const
     {
         return double2(r * x_, r * y_);
     }
-    double modulus() const
+    float modulus() const
     {
         return sqrt(x_ * x_ + y_ * y_);
     }
     double2 unit_vector() const
     {
         const auto mod = modulus();
-        if (abs(mod) < std::numeric_limits<double>::epsilon())
+        if (abs(mod) < std::numeric_limits<float>::epsilon())
             return double2();
         return double2(x_ / mod, y_ / mod);
     }
-    double distance(const double2& r) const
+    float distance(const double2& r) const
     {
         const auto d = *this - r;
         const auto x2 = d.x_ * d.x_;
@@ -66,22 +66,22 @@ public:
         return sqrt(x2 + y2);
     }
 
-    double x() const { return x_; }
-    double y() const { return y_; }
+    float x() const { return x_; }
+    float y() const { return y_; }
 private:
-    friend double2 operator*(double f, const double2& r);
-    friend double dot_product(const double2& a, const double2& b);
+    friend double2 operator*(float f, const double2& r);
+    friend float dot_product(const double2& a, const double2& b);
 
-    double x_;
-    double y_;
+    float x_;
+    float y_;
 };
 
-double2 operator*(double f, const double2& r)
+double2 operator*(float f, const double2& r)
 {
     return double2(f * r.x_, f * r.y_);
 }
 
-double dot_product(const double2& a, const double2& b)
+float dot_product(const double2& a, const double2& b)
 {
     return a.x_ * b.x_ + a.y_ * b.y_;
 }
