+++ physics/mass.h
@@ -6,21 +6,21 @@
 class mass
 {
 public:
-    mass(double m, double r) : m_(m), r_(r)
+    mass(float m, float r) : m_(m), r_(r)
     {
     }
 
-    double m() const
+    float m() const
     {
         return m_;
     }
 
-    double r() const
+    float r() const
     {
         return r_;
     }
 
 private:
-    double m_;                  //  kg
-    double r_;                  //  m
+    float m_;                  //  kg
+    float r_;                  //  m
 };
